+++ barman/clients/cloud_backup_list.py
@@ -21,7 +21,8 @@ import logging
 from contextlib import closing
 
 import barman
-from barman.cloud import S3CloudInterface, CloudBackupCatalog, configure_logging
+from barman.cloud import CloudBackupCatalog, configure_logging
+from barman.cloud_providers import S3CloudInterface
 from barman.infofile import BackupInfo
 from barman.utils import force_str
