+++ barman/clients/cloud_walarchive.py
@@ -26,7 +26,8 @@ from contextlib import closing
 from io import BytesIO
 
 import barman
-from barman.cloud import S3CloudInterface, configure_logging
+from barman.cloud import configure_logging
+from barman.cloud_providers import S3CloudInterface
 from barman.utils import force_str
 from barman.xlog import hash_dir, is_any_xlog_file
